Add base class for submitters
